RP special: SPDX experiment
Two commits:
package: Generate srclist and filelics json files
Change do_package to:

a) Save the debug object -> source file mapping
information (in TEMPDBGSRCMAPPING)

b) Save a .srclist file containing that information
(in json format)

c) Perform a grep of all the sources in the build
of this recipe looking for SPDX license headers

d) Save out a file listing the SPDX headers found
for each file

package: Add license computation experiment

This code creates a list of SPDX headers found for
the sources that make up a given set of binaries that
make up an individual package.

This is then compared with the license field of the
given package containing those binaries.

Signed-off-by: student@e450-7 <student@e450-7.res.training>